extracted RecyclerViewItem in nested example
