+++ AndroidBoundExample-support/src/main/java/solutions/alterego/androidbound/example/support/nestedrvs/viewmodel/MainNestedViewModel.java
@@ -6,86 +6,9 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import lombok.experimental.Accessors;
 import solutions.alterego.androidbound.ViewModel;
 import solutions.alterego.androidbound.example.support.R;
 
-public class MainNestedViewModel {
-
-    public List<Object> getDataItems() {
-        return this.mDataItems;
-    }
-
-    @Accessors(prefix = "m")
-    public static class RecyclerViewItem {
-
-        private String mName;
-
-        private String mImageUrl;
-
-        public RecyclerViewItem(String name, String imageUrl) {
-            mName = name;
-            mImageUrl = imageUrl;
-        }
-
-        public String getName() {
-            return this.mName;
-        }
-
-        public String getImageUrl() {
-            return this.mImageUrl;
-        }
-
-        public void setName(String mName) {
-            this.mName = mName;
-        }
-
-        public void setImageUrl(String mImageUrl) {
-            this.mImageUrl = mImageUrl;
-        }
-
-        public boolean equals(Object o) {
-            if (o == this) {
-                return true;
-            }
-            if (!(o instanceof RecyclerViewItem)) {
-                return false;
-            }
-            final RecyclerViewItem other = (RecyclerViewItem) o;
-            if (!other.canEqual((Object) this)) {
-                return false;
-            }
-            final Object this$mName = this.getName();
-            final Object other$mName = other.getName();
-            if (this$mName == null ? other$mName != null : !this$mName.equals(other$mName)) {
-                return false;
-            }
-            final Object this$mImageUrl = this.getImageUrl();
-            final Object other$mImageUrl = other.getImageUrl();
-            if (this$mImageUrl == null ? other$mImageUrl != null : !this$mImageUrl.equals(other$mImageUrl)) {
-                return false;
-            }
-            return true;
-        }
-
-        public int hashCode() {
-            final int PRIME = 59;
-            int result = 1;
-            final Object $mName = this.getName();
-            result = result * PRIME + ($mName == null ? 43 : $mName.hashCode());
-            final Object $mImageUrl = this.getImageUrl();
-            result = result * PRIME + ($mImageUrl == null ? 43 : $mImageUrl.hashCode());
-            return result;
-        }
-
-        protected boolean canEqual(Object other) {
-            return other instanceof RecyclerViewItem;
-        }
-
-        public String toString() {
-            return "MainNestedViewModel.RecyclerViewItem(mName=" + this.getName() + ", mImageUrl=" + this.getImageUrl() + ")";
-        }
-    }
 public class MainNestedViewModel extends ViewModel {
 
     private List<Object> mDataItems;
@@ -101,8 +24,12 @@ public class MainNestedViewModel extends ViewModel {
     public Map<Class<?>, Integer> getObjectTemplates() {
         Map<Class<?>, Integer> map = new HashMap<Class<?>, Integer>();
         map.put(NestedViewModel.class, R.layout.nested_recycler_view);
-        map.put(MainNestedViewModel.RecyclerViewItem.class, R.layout.activity_paginated_rv_item);
+        map.put(RecyclerViewItem.class, R.layout.activity_paginated_rv_item);
         return map;
+    }
+
+    public List<Object> getDataItems() {
+        return mDataItems;
     }
 
 }
