+++ AndroidBoundExample-support/src/main/java/solutions/alterego/androidbound/example/support/nestedrvs/viewmodel/NestedViewModel.java
@@ -8,12 +8,12 @@ import solutions.alterego.androidbound.ViewModel;
 
 public class NestedViewModel extends ViewModel {
 
-    private List<MainNestedViewModel.RecyclerViewItem> mNestedDataSet;
+    private List<RecyclerViewItem> mNestedDataSet;
 
     public NestedViewModel() {
-        mNestedDataSet = new ArrayList<MainNestedViewModel.RecyclerViewItem>();
+        mNestedDataSet = new ArrayList<RecyclerViewItem>();
         for (int i = 0; i < 20; i++) {
-            mNestedDataSet.add(new MainNestedViewModel.RecyclerViewItem("nested item " + i,
+            mNestedDataSet.add(new RecyclerViewItem("nested item " + i,
                     "https://www.google.co.uk/images/branding/googlelogo/1x/googlelogo_color_272x92dp.png"));
         }
     }
@@ -22,7 +22,7 @@ public class NestedViewModel extends ViewModel {
         return true;
     }
 
-    public List<MainNestedViewModel.RecyclerViewItem> getNestedDataSet() {
-        return this.mNestedDataSet;
+    public List<RecyclerViewItem> getNestedDataSet() {
+        return mNestedDataSet;
     }
 }
